+++ lib/packages/flutter_typeahead-4.3.7/lib/src/material/suggestions_box/suggestions_box.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_typeahead/src/material/field/typeahead_field.dart';
+import 'package:flutter_typeahead_mm/src/material/field/typeahead_field.dart';
 
 class SuggestionsBox {
   static const int waitMetricsTimeoutMillis = 1000;
@@ -21,11 +21,11 @@ class SuggestionsBox {
   late double directionUpOffset;
 
   SuggestionsBox(
-      this.context,
-      this.direction,
-      this.autoFlipDirection,
-      this.autoFlipListDirection,
-      ) : desiredDirection = direction;
+    this.context,
+    this.direction,
+    this.autoFlipDirection,
+    this.autoFlipListDirection,
+  ) : desiredDirection = direction;
 
   void open() {
     if (this.isOpened) return;
@@ -131,7 +131,7 @@ class SuggestionsBox {
     if (maxHDesired >= minOverlaySpace || !autoFlipDirection) {
       direction = desiredDirection;
       // Sometimes textBoxAbsY is NaN, so we need to check for that
-      if(!maxHDesired.isNaN) {
+      if (!maxHDesired.isNaN) {
         maxHeight = maxHDesired;
       }
     } else {
@@ -145,7 +145,7 @@ class SuggestionsBox {
         direction = flipped;
 
         // Not sure if this is needed, but it's here just in case
-        if(!maxHFlipped.isNaN) {
+        if (!maxHFlipped.isNaN) {
           maxHeight = maxHFlipped;
         }
       }
@@ -164,9 +164,9 @@ class SuggestionsBox {
       double textBoxAbsY) {
     return direction == AxisDirection.down
         ? _calculateMaxHeightDown(box, widget, windowHeight, rootMediaQuery,
-        keyboardHeight, textBoxAbsY)
+            keyboardHeight, textBoxAbsY)
         : _calculateMaxHeightUp(box, widget, windowHeight, rootMediaQuery,
-        keyboardHeight, textBoxAbsY);
+            keyboardHeight, textBoxAbsY);
   }
 
   double _calculateMaxHeightDown(
@@ -179,7 +179,7 @@ class SuggestionsBox {
     // unsafe area, ie: iPhone X 'home button'
     // keyboardHeight includes unsafeAreaHeight, if keyboard is showing, set to 0
     double unsafeAreaHeight =
-    keyboardHeight == 0 ? rootMediaQuery.data.padding.bottom : 0;
+        keyboardHeight == 0 ? rootMediaQuery.data.padding.bottom : 0;
 
     return windowHeight -
         keyboardHeight -
@@ -208,11 +208,11 @@ class SuggestionsBox {
 
     return textBoxAbsY > keyboardAbsY
         ? keyboardAbsY -
-        unsafeAreaHeight -
-        2 * widget.suggestionsBoxVerticalOffset
+            unsafeAreaHeight -
+            2 * widget.suggestionsBoxVerticalOffset
         : textBoxAbsY -
-        unsafeAreaHeight -
-        2 * widget.suggestionsBoxVerticalOffset;
+            unsafeAreaHeight -
+            2 * widget.suggestionsBoxVerticalOffset;
   }
 
   Future<void> onChangeMetrics() async {