+++ lib/src/views/explore/models/space_model.dart
@@ -1,3 +1,4 @@
+// ignore: depend_on_referenced_packages
 import 'package:json_annotation/json_annotation.dart';
 
 part 'space_model.g.dart';