+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 ## 1.0.0
 
 Initial stable release v1.0.0
+
+## 1.1.0
+
+Added search in participants list
+Added member rights